fix: make code clean
