+++ src/components/Reactions/getUniqueEmojiCodes.js
@@ -1,0 +1,23 @@
+import _ from 'underscore';
+import getPreferredEmojiCode from './getPreferredEmojiCode';
+
+/**
+ * Given an emoji object and a list of senders it will return an
+ * array of emoji codes, that represents all used variations of the
+ * emoji.
+ * @param {{ name: string, code: string, types: string[] }} emoji
+ * @param {Array} users
+ * @return {string[]}
+ * */
+export default function  getUniqueEmojiCodes(emoji, users) {
+    const emojiCodes = [];
+    _.forEach(users, (user) => {
+        const emojiCode = getPreferredEmojiCode(emoji, user.skinTone);
+
+        if (emojiCode && !emojiCodes.includes(emojiCode)) {
+            emojiCodes.push(emojiCode);
+        }
+    });
+    return emojiCodes;
+};
+
